requirement.txt
python-dotenv
tqdm
opencv-python
tensorflow_addons
tensorflow
matplotlib
keras-rectified-adam
numpy
sklearn
